Update state property to include all team fields
Mount $state propriety to include all existing fields of the Team model 
in order to allow the implementation of additional fields for this Model. 
That way, the new fields can be added to the teams table, and implement them into the customizable views and actions to update-team-name.
The original way allows only the use of "name" key in the state array.