Fix: jetstream browser sessions ignoring session connection in config
When using Jetstream with the Livewire stack, It is impossible to use a second database connection specifically for sessions. As the component ignores the session.connection config value.

Could also be fixed by allowing livewire components to be published from vendor folder. If It is possible it's an oversight on my part.