Validation message for existing invitation
When an email has already been invited but the same email address also gets invited to another team, but not yet accepted the past invitation the error message implies that the email is already added to the current team. So the error message could be a bit confusing so think it's better to inform the email already has a pending invitation.