Changed the column type of profile_photo_path to string(2048)
As the TEXT type on some DBs, creates an column with a limit of 2/4 GB.

This should increase compatibly with other DBs, with no impact on the migration it self. But we DO limit the column length, which is a good thing for this column use case. As a good practice URI should not be longer then 2K.

There is no mention of max URI length in the RFC3986, but there are limitation imposed by the software like the browsers them self.