Allow team invitation table to use connection out of the box...
A project I work on uses Spatie's multi-tenancy package with a landlord database and multiple tenant databases. In order to guarantee that the correct connection is used, I need to change it in this file (`Rule::unique('my_connection.team_invitations')`).

This change should allow the correct connection/table in all (?) cases without having to change anything here.

I'm not convinced this is the best way to resolve the problem.  I'd love feedback or, if it's not clear, any questions about why this might be necessary.

Thanks.  🤓