Better handle sitatuations where a user's role on a team is undefined
In some situations (especially when testing using database factories), the "role" column on the team_user table may be left NULL. As a result, attempting to check the user's permissions based on their role causes an error, as `Jetstream::findRole()` requires a string.

For example:

```php
use App\Models\Team;
use App\Models\User;

$team = Team::factory()
	->with(User::factory())
	->create();

$team->users->first()->hasTeamPermission($team, 'read');
```

This commit ensures that `Jetstream::findRole()` is only called if we have a corresponding role, preventing this error. Similarly, `teamPermissions()` will no longer cause a "ErrorException: Attempt to read property "permissions" on null" exception when the role is undefined.